projects/RSGPT/requirements.txt
langchain
openai
langchain_openai
langchain_core
tdqm
python-dotenv